requirements.txt
aiofiles==0.6.0
alembic==1.5.3
astroid==2.4.2
asyncpg==0.22.0
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
databases==0.4.1
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
importlib-metadata==2.1.1
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
numpy @ file:///C:/Users/Aleksey/Envs/numpy-1.19.5%2Bvanilla-cp38-cp38-win_amd64.whl
opencv-python @ file:///C:/Users/Aleksey/Envs/opencv_python-4.4.0-cp38-cp38-win_amd64.whl
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
shellingham==1.4.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
urllib3==1.26.3
uvicorn==0.13.3
watchgod==0.6
websockets==8.1
wrapt==1.12.1
zipp==3.4.0
